feat: BankSystem 관련 폴더링
